fix(api): Create method not returning response
- Artists, Audios and Images now return response data on Create method
- Moved soundcloud preview method
