+++ src/stores/api/crud/audios.ts
@@ -16,8 +16,10 @@ export async function getAudio(audio_id: string): Promise<Audio> {
   return res.data as Audio;
 }
 
-export async function createAudio(file_id: string) {
+export async function createAudio(file_id: string): Promise<Audio> {
   const res = await axios.post(`${API_LOCATION}/db/audios/`, {
     file: file_id,
   });
+
+  return res.data as Audio;
 }
