+++ src/stores/api/import/soundcloud.ts
@@ -1,6 +1,15 @@
 import axios from "axios";
 import { API_LOCATION, type Music } from "..";
 
+export type SoundcloudTrack = {
+  title: string;
+  url: string;
+  thumbnail: string;
+  duration: { secs: number };
+  media: string;
+  client_id: string;
+};
+
 export type SoundcloudImport = {
   title: string;
   image_id: string;
@@ -13,4 +22,10 @@ export type SoundcloudImport = {
 export async function soundcloudImport(data: SoundcloudImport): Promise<Music> {
   const res = await axios.post(`${API_LOCATION}/soundcloud/import/`, data);
   return res.data as Music;
+}
+
+export async function previewFromUrl(url: string) {
+  const url_encoded = encodeURIComponent(url);
+  const res = await fetch(`${API_LOCATION}/soundcloud/metadata/${url_encoded}`);
+  return (await res.json()) as SoundcloudTrack;
 }
